user_management/requirements.txt
requests
django
djangorestframework
django-cors-headers
psycopg2-binary
psycopg2
psycopg
PyJWT
Pillow
